+++ hail/hail/src/is/hail/services/oauth2.scala
@@ -90,7 +90,7 @@ object oauth2 {
       }
 
     private[this] def isExpired: Boolean =
-      token == null || OffsetDateTime.now.plusMinutes(AzureTokenRefreshMinutes).isBefore(
+      token == null || OffsetDateTime.now.plusMinutes(AzureTokenRefreshMinutes.toLong).isBefore(
         token.getExpiresAt
       )
   }
